shipping_microservice/requirements.txt
Flask==2.2.3
python-dotenv==0.20.0
mysql-connector-python==8.0.31
flask-sqlalchemy
flask-restful==0.3.9
flask-marshmallow==0.14.0